Fix errant write in Daira code.
There is an errant write to ROM in the Daira routines.  In the FDS
version of the game, this address was almost certainly in RAM, but in
the USA cartridge release, the target address is ROM.  This has the side
effect of writing to the mapper hardware, usually with bad results.

Signed-off-by: Chris Frantz <frantzcj@gmail.com>