Fix door unlock sound
In the vanilla game, the "door unlock" and "boss death" sounds are the same
sound effect.  The vanilla game cancels the "boss death" sound after 17
frames to make the door unlock sound.

I've reassigned one of the unused slots for "boss death" and created a
shortened effect for "door unlock".

Signed-off-by: Chris Frantz <frantzcj@gmail.com>