Exchange the fairy and file delete sounds
The game utilizes the sound effect number of the fairy/raft to compute
the starting position of the raft animation on the overworld.

In vanilla, the SFX number is $C0, which gets shifted left to compute
the value $80 (screen midpoint) as the start position of the animation.
Therefore, I'm making the new raft sound as SFX number $80 and `nop`ing
out the shift-left instruction.

Signed-off-by: Chris Frantz <frantzcj@gmail.com>